Move Stats::submit to solana-runtime
