+++ program-runtime/src/loaded_programs.rs
@@ -3,7 +3,7 @@ use {
         invoke_context::{BuiltinFunctionWithContext, InvokeContext},
         timings::ExecuteDetailsTimings,
     },
-    log::{debug, error, log_enabled, trace},
+    log::error,
     percentage::PercentageInteger,
     rand::{thread_rng, Rng},
     solana_measure::measure::Measure,
@@ -203,58 +203,6 @@ pub struct Stats {
 }
 
 impl Stats {
-    /// Logs the measurement values
-    pub fn submit(&self, slot: Slot) {
-        let hits = self.hits.load(Ordering::Relaxed);
-        let misses = self.misses.load(Ordering::Relaxed);
-        let evictions: u64 = self.evictions.values().sum();
-        let reloads = self.reloads.load(Ordering::Relaxed);
-        let insertions = self.insertions.load(Ordering::Relaxed);
-        let lost_insertions = self.lost_insertions.load(Ordering::Relaxed);
-        let replacements = self.replacements.load(Ordering::Relaxed);
-        let one_hit_wonders = self.one_hit_wonders.load(Ordering::Relaxed);
-        let prunes_orphan = self.prunes_orphan.load(Ordering::Relaxed);
-        let prunes_environment = self.prunes_environment.load(Ordering::Relaxed);
-        let empty_entries = self.empty_entries.load(Ordering::Relaxed);
-        datapoint_info!(
-            "loaded-programs-cache-stats",
-            ("slot", slot, i64),
-            ("hits", hits, i64),
-            ("misses", misses, i64),
-            ("evictions", evictions, i64),
-            ("reloads", reloads, i64),
-            ("insertions", insertions, i64),
-            ("lost_insertions", lost_insertions, i64),
-            ("replace_entry", replacements, i64),
-            ("one_hit_wonders", one_hit_wonders, i64),
-            ("prunes_orphan", prunes_orphan, i64),
-            ("prunes_environment", prunes_environment, i64),
-            ("empty_entries", empty_entries, i64),
-        );
-        debug!(
-            "Loaded Programs Cache Stats -- Hits: {}, Misses: {}, Evictions: {}, Reloads: {}, Insertions: {} Lost-Insertions: {}, Replacements: {}, One-Hit-Wonders: {}, Prunes-Orphan: {}, Prunes-Environment: {}, Empty: {}",
-            hits, misses, evictions, reloads, insertions, lost_insertions, replacements, one_hit_wonders, prunes_orphan, prunes_environment, empty_entries
-        );
-        if log_enabled!(log::Level::Trace) && !self.evictions.is_empty() {
-            let mut evictions = self.evictions.iter().collect::<Vec<_>>();
-            evictions.sort_by_key(|e| e.1);
-            let evictions = evictions
-                .into_iter()
-                .rev()
-                .map(|(program_id, evictions)| {
-                    format!("  {:<44}  {}", program_id.to_string(), evictions)
-                })
-                .collect::<Vec<_>>();
-            let evictions = evictions.join("\n");
-            trace!(
-                "Eviction Details:\n  {:<44}  {}\n{}",
-                "Program",
-                "Count",
-                evictions
-            );
-        }
-    }
-
     pub fn reset(&mut self) {
         *self = Stats::default();
     }
